Cycles : Move output driver out of subdirectory
Unlike the Arnold driver, this isn't a separate plugin that is loaded by the renderer. It's just a class that we use directly in `CyclesRenderer`, so it can be an entirely private component of `IECoreCyclesPreview`.
